Replace "= size_t(-1)" with " -1" for size_t

+++ include/OOBase/BTree.h
@@ -717,7 +717,7 @@ namespace OOBase
 					else
 						end = mid;
 				}
-				return size_t(-1);
+				return -1;
 			}
 
 			bool insert_i(BTreeImpl<K,V,Compare,B,Allocator>* tree, const Pair<K,V>& value, size_t parent_pos)
